feat/pedir_todas_materias_primas + refactor
Se crea funcion pedir_todas_materias_primas que pide todo.
Se mueven funciones de api bodega desde application_helper a api_bodega_helper
Se crea api_bodega_helper para interactuar con la api bodega
Se crea grupos_helper para interactuar con los grupos.

Se agrega ruta /pedir_todo que llama a la funcion pedir_todas_materias_primas
