+++ app/controllers/application_controller.rb
@@ -4,6 +4,7 @@ require 'net/ftp'
 require 'application_helper'
 require 'variables_helper'
 
+
 require 'active_support/core_ext/hash'
 require 'date'
 
@@ -12,6 +13,8 @@ class ApplicationController < ActionController::Base
 	include ApplicationHelper
 	include OcHelper
 	include VariablesHelper
+	include ApiBodegaHelper
+	include GruposHelper
 
 	def start
 		# Funcion que llama a /inventories
