+++ config/routes.rb
@@ -10,4 +10,7 @@ get '/totalproducts', to: 'inventories#total_products'
 get '/getskustock', to: 'inventories#sku_stock'
 get '/checkin_init', to: 'inventories#init_check_inventory'
 post '/documents/:order_id/notification', to: 'documents#notificaciones'
+
+get '/pedir_todo', to: 'inventories#pedir_todo'
+
 end
